SKALE-3082 Update most of the tests
